pyproject.toml
[project]
name = "budgetplanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "streamlit>=1.45.1",
]
